requirements.txt
scikit-learn==1.2.1
tqdm==4.65.0
Cython==0.29.33
matplotlib==3.7.0
numpy==1.24.2
pycocotools==2.0.6
Pillow==9.4.0
editdistance==0.6.2
scikit-image==0.20.0
scipy==1.10.1
PyMuPDF==1.21.1
pandas==1.5.3